requirements.txt
python-dotenv==1.2.1
pytest==8.4.2
requests==2.32.5